update tests and change to typescript
