+++ fields.md
@@ -2,7 +2,7 @@
 
 This is a comprehensive list of possible fields that may appear in incoming request data from Overseerr/Jellyseerr and can be used in filters.
 
-Example values for each field can be found in [filters.test.js](https://github.com/varthe/Redirecterr/blob/main/filters.test.js)
+Example values for each field can be found in [filters.test.ts](https://github.com/varthe/Redirecterr/blob/main/filters.test.ts)
 
 - `requestedBy_email`
 - `requestedBy_username`
@@ -32,8 +32,7 @@ Example values for each field can be found in [filters.test.js](https://github.c
 - `spokenLanguages`
 - `productionCompanies`
 - `productionCountries`
-- `networks` 
+- `networks`
 - `inProduction`
 - `numberOfSeasons` / `numberOfEpisodes`
 - `contentRatings`
-
